requirements.txt
fastapi
python-dotenv
uvicorn
requests
python-dotenv
